fix local variable mutexes
Problem: local variable mutexes are thread specific - each thread
will create its own mutex, which in turn won't lock anything.cpp17

Solution: make local variable mutexes static - one static will shared
by all threads
