+++ basecom.hpp
@@ -33,7 +33,7 @@
 #include <unistd.h>
 
 #include <epoll.hpp>
-#include <logger.hpp>
+#include <log/logger.hpp>
 
 static const char str_unknown[] = "unknown";
 static const char str_getsockname[] = "getsockname";
