+++ baseproxy.hpp
@@ -27,7 +27,7 @@
 #include <unistd.h>
 #include <vector>
 
-#include <logger.hpp>
+#include <log/logger.hpp>
 #include <hostcx.hpp>
 
 /*
