+++ common/internet.cpp
@@ -1,5 +1,5 @@
 #include <internet.hpp>
-#include <logger.hpp>
+#include <log/logger.hpp>
 
 namespace inet {
