+++ epoll.hpp
@@ -16,10 +16,9 @@
 #include <arpa/inet.h>
 #include <netdb.h>
 #include <fcntl.h>
-#include <time.h>
 #include <unistd.h>
 
-#include <logger.hpp>
+#include <log/logan.hpp>
 
 #define EPOLLER_MAX_EVENTS 50
 #define HANDLER_FENCE 0xcaba1a
