+++ sobject.hpp
@@ -24,7 +24,7 @@
 
 #include <time.h>
 
-#include <logger.hpp>
+#include <log/logger.hpp>
 #include <ptr_cache.hpp>
 #include <display.hpp>
