+++ tcpcom.hpp
@@ -32,7 +32,7 @@
 
 #include <ctime>
 
-#include <logger.hpp>
+#include <log/logger.hpp>
 #include <basecom.hpp>
 #include <display.hpp>
